+++ .gitignore
@@ -237,3 +237,4 @@ src/main/resources/application.properties
 .env
 /build/
 docker-compose.yml
+/build
