--- src/main/java/com/tools/seoultech/timoproject/domain/Post.java
@@ -1,25 +1,0 @@
-package com.tools.seoultech.timoproject.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Getter
-@ToString
-public class Post extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(length=100, nullable=false)
-    private String title;
-
-    @Column(length=1500, nullable=false)
-    private String content;
-
-    @Column(length=50, nullable=false)
-    private String writer;
-}
