--- src/main/java/com/tools/seoultech/timoproject/dto/APIDataResponse.java
@@ -1,23 +1,0 @@
-package com.tools.seoultech.timoproject.dto;
-
-import com.tools.seoultech.timoproject.constant.ErrorCode;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-public class APIDataResponse<T> extends APIErrorResponse{
-    private final T data;
-
-    private APIDataResponse(T data) {
-        super(true, ErrorCode.OK.getCode(), ErrorCode.OK.getMessage());
-        this.data = data;
-    }
-
-    public static <T> APIDataResponse<T> of(T data) {
-        return new APIDataResponse<>(data);
-    }
-
-    public static <T> APIDataResponse<T> empty() {
-        return new APIDataResponse<>(null);
-    }
-}
