--- src/main/java/com/tools/seoultech/timoproject/dto/PostDTO.java
@@ -1,21 +1,0 @@
-package com.tools.seoultech.timoproject.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-public class PostDTO {
-    private Long id;
-    private String title;
-    private String content;
-    private String writer;
-    private LocalDateTime regDate, modDate;
-}
-
