+++ src/main/java/com/tools/seoultech/timoproject/post/domain/Post.java
@@ -1,0 +1,26 @@
+package com.tools.seoultech.timoproject.post.domain;
+
+import com.tools.seoultech.timoproject.global.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Getter
+@ToString
+public class Post extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy= GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(length=100, nullable=false)
+    private String title;
+
+    @Column(length=1500, nullable=false)
+    private String content;
+
+    @Column(length=50, nullable=false)
+    private String writer;
+}
