+++ src/main/java/com/tools/seoultech/timoproject/post/repository/PostRepository.java
@@ -1,0 +1,11 @@
+package com.tools.seoultech.timoproject.post.repository;
+
+import com.tools.seoultech.timoproject.post.domain.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.querydsl.QuerydslPredicateExecutor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostRepository extends JpaRepository<Post, Long>, QuerydslPredicateExecutor<Post> {
+
+}
