--- src/main/java/com/tools/seoultech/timoproject/repository/PostRepository.java
@@ -1,12 +1,0 @@
-package com.tools.seoultech.timoproject.repository;
-
-import com.tools.seoultech.timoproject.domain.Post;
-import jakarta.persistence.Id;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.querydsl.QuerydslPredicateExecutor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PostRepository extends JpaRepository<Post, Long>, QuerydslPredicateExecutor<Post> {
-
-}
