+++ src/test/java/com/tools/seoultech/timoproject/repository/PostRepositoryTest.java
@@ -1,14 +1,13 @@
 package com.tools.seoultech.timoproject.repository;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.tools.seoultech.timoproject.domain.Post;
+import com.tools.seoultech.timoproject.post.domain.Post;
+import com.tools.seoultech.timoproject.post.repository.PostRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import java.util.stream.IntStream;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 class PostRepositoryTest {
